% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEExperimentHub.R
\name{iSEEExperimentHub}
\alias{iSEEExperimentHub}
\title{iSEEExperimentHub app}
\usage{
iSEEExperimentHub(ehub)
}
\arguments{
\item{ehub}{An instance of \code{\link[=ExperimentHub]{ExperimentHub()}}.}
}
\value{
An \code{\link[=iSEE]{iSEE()}} app with a custom landing page interfacing with the
Bioconductor \code{\link[=ExperimentHub]{ExperimentHub()}}.
}
\description{
iSEEExperimentHub app
}
\examples{
library(ExperimentHub)
ehub <- ExperimentHub()

app <- iSEEExperimentHub(ehub)

if (interactive()) {
  shiny::runApp(app, port = 1234)
}
}
