% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEhub.R
\name{iSEEhub}
\alias{iSEEhub}
\title{iSEEhub app}
\usage{
iSEEhub(ehub)
}
\arguments{
\item{ehub}{An \code{\link[=ExperimentHub]{ExperimentHub()}} object.}
}
\value{
An \code{\link[=iSEE]{iSEE()}} app with a custom landing page interfacing with \code{ehub}.
}
\description{
iSEEhub app
}
\examples{
library(ExperimentHub)
ehub <- ExperimentHub()

app <- iSEEhub(ehub)

if (interactive()) {
  shiny::runApp(app, port = 1234)
}
}
